backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.9.2
python-dotenv==1.0.1
openai>=1.43.0
python-dateutil==2.9.0.post0
pytz==2024.1
orjson==3.10.7
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
